Move the respective properties of the `DefaultObjectAccessPermission` into the companion object and extract fold method.
